+++ app/src/main/java/org/sopt/stamp/data/mapper/MissionsMapper.kt
@@ -1,11 +1,11 @@
 package org.sopt.stamp.data.mapper // ktlint-disable filename
 
-import org.sopt.stamp.data.model.MissionData
+import org.sopt.stamp.data.remote.model.MissionsData
 import org.sopt.stamp.domain.model.Mission
 
-internal fun List<MissionData>.toDomain(): List<Mission> = this.map { it.toDomain() }
+internal fun List<MissionsData.MissionData>.toDomain(): List<Mission> = this.map { it.toDomain() }
 
-internal fun MissionData.toDomain(): Mission = Mission(
+internal fun MissionsData.MissionData.toDomain(): Mission = Mission(
     id = this.id,
     title = this.title,
     level = this.level,
